Temporary fix for search results with only one hit
If a search just return one hit, send the result page with only
one hit instead of redirecting to the file unless we know the
anchor in that page (Only true for TermQyery). If you run with
projects enabled most queries will be BooleanQuery. This should
be improved to determine the correct anchor and redirect to the
correct page.