Add -r command line option to set if we should ignore remote SCM systems
The current implementation will try to download SCM history as long as
the underlying SCM implementation allows remote access. This might be
a time-consuming task, so it should be possible to specify that such
repositories should be ignored. This command-line option will toggle a flag
that the various SCM implementations may check.