Removed createCache() and directory history parser from RazorRepository
The cache will probably not speed up history access for Razor
repositories, as the history is already available locally, so remove
it to simplify the code. Also removed RazorDirectoryHistoryParser
since there's no efficient way to obtain the full history of a
directory in a Razor repository, and the current implementation would
always return a single entry only.