Bug #3624: History cache should be stored in Java DB
Got rid of the system property to enable JDBCHistoryCache and added a
field in the Configuration class instead. No parameter to enable it
exists yet, so you'll need to add the following to the configuration
file to enable it:

<void property="historyCacheInDB">
  <boolean>true</boolean>
</void>

You'll also need to add derby.jar to the classpath to get it to work.