Manually generate identifiers for some tables.
When running initial indexing on many large code bases in
parallel, there may be severe lock contention on the system table
that holds the state of the id generators. Maintaining a counter
manually in an AtomicInteger for the hot tables is cheaper and
scales better than using SQL auto-generated identifiers.